+++ NEWS.md
@@ -62,7 +62,7 @@
 
 ## Bug fixes
 
--   Autoregression via `ar()` is removed, now giving an error for Poisson models. Support may be added in the future when handling of zeroes, latent-scale vs. observed-scale AR, etc. is thought through.
+-   Autoregression via `ar()` is temporarily restricted to `gaussian(link = "identity")` models while generalized autoregression is implemented explicitly.
 
 -   The quantiles for `fitted()` and `predict()` for varying-changepoint models ignored the varying level - they were identical across levels.